Added common code for initializing the health API instead of code duplication
