+++ stackstate_checks_base/stackstate_checks/base/checks/__init__.py
@@ -10,7 +10,7 @@ except ImportError:
     from .base import AgentCheck, TopologyInstance, StackPackInstance, AgentIntegrationInstance
     from .v2 import AgentCheckV2, StatefulMixin, TransactionalMixin, StatefulAgentCheck, TransactionalAgentCheck, \
         CheckResponse
-    from ..utils.health_api import HealthStream, HealthStreamUrn, Health
+    from ..utils.health_api import HealthStream, HealthStreamUrn, Health, HealthApiCommon
     from .network import NetworkCheck, Status, EventType
 
 __all__ = [
@@ -23,6 +23,7 @@ __all__ = [
     'EventType',
     'HealthStream',
     'HealthStreamUrn',
+    'HealthApiCommon',
     'Health',
     'AgentCheckV2',
     'StatefulMixin',
