+++ stackstate_checks_base/stackstate_checks/base/checks/v2/mixins/health.py
@@ -1,10 +1,9 @@
 from typing import Any, Optional
 from .mixins import CheckMixin
-from ..types import InstanceType
-from ....utils.health_api import HealthApi, HealthStream
+from ....utils.health_api import HealthApi, HealthApiCommon
 
 
-class HealthMixin(CheckMixin):
+class HealthMixin(CheckMixin, HealthApiCommon):
     """
     HealthMixin extends the Agent Check base with the Health API allowing Agent Checks to start submitting health data.
     """
@@ -23,16 +22,6 @@ class HealthMixin(CheckMixin):
         # self.instance =
         self.health = None  # type: Optional[HealthApi]
 
-    def get_health_stream(self, instance):  # type: (InstanceType) -> Optional[HealthStream]
-        """
-        Integration checks can override this if they want to be producing a health stream. Defining this will
-        enable self.health() calls.
-
-        @param instance:
-        @return: a class extending HealthStream
-        """
-        return None
-
     def setup(self):  # type: () -> None
         """
         setup is used to initialize the HealthApi and set the health variable to be used in checks.
@@ -41,20 +30,5 @@ class HealthMixin(CheckMixin):
         """
         super(HealthMixin, self).setup()
 
-        if self.health is not None:
-            return None
-
-        stream_spec = self.get_health_stream(self._get_instance_schema(self.instance))
-        if stream_spec:
-            # collection_interval should always be set by the agent
-            collection_interval = self.instance['collection_interval']
-            repeat_interval_seconds = stream_spec.repeat_interval_seconds or collection_interval
-            expiry_seconds = stream_spec.expiry_seconds
-            # Only apply a default expiration when we are using sub_streams
-            if expiry_seconds is None:
-                if stream_spec.sub_stream != "":
-                    expiry_seconds = repeat_interval_seconds * 4
-                else:
-                    # Explicitly disable expiry setting it to 0
-                    expiry_seconds = 0
-            self.health = HealthApi(self, stream_spec, expiry_seconds, repeat_interval_seconds)
+        # Initialize the health api
+        HealthApiCommon._init_health_api(self)
